updated files

+++ lib/functions.php
@@ -144,6 +144,7 @@ function savePoster($link, $img, $code)
     if (file_exists($seriefile) || al_is_serie($path_parts['dirname']))
         copy($img,'../images/poster_small/s-'.$code.'.jpg');
     if ($POSTER_WITH_VIDEO == TRUE) {
+        // TODO : for series, only copy the first episode and make link to it for all other episodes
         copy('../images/poster_small/'.$code.'.jpg',$path_parts['dirname'] . '/' . $path_parts['filename'] . '.jpg');   
     }
     
@@ -285,20 +286,57 @@ function is_serie($SERIES_DIR){
 }
 */
 
-function banner_serie(){
-    global $DELETED_WORDS;
-	$allo = new AlloCine();
-	$infos = explode('/',urldecode($_GET['rep']));
-	if(count($infos)>= 4){ // $$AL$$ TODO 
-		$name = $infos[count($infos) - 1];
-        debug("search serie " . $name);
-		$recherche = $allo->serieSearch(keywordsAdapt($name,$DELETED_WORDS,1));//recherche serie
-		if (!empty($recherche['code'])){
-		    $id=$recherche['code'];
-		    $serie = $allo->serieInfos($id);
-		    return $serie['topBanner'];
-		}
-	}
+function get_serie_code_infos($dir) {
+    $sql = "SELECT * FROM series_id WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."'";
+	$req = mysql_query($sql) or die ('Erreur SQL'.$sql.mysql_error());
+    if (mysql_num_rows($req) == 1) {
+        $data = mysql_fetch_array($req);
+        return $data;
+    }
+    $data = array();
+    return $data;
+}
+
+function set_serie_code($dir, $code) {
+    global $SERIES_DATABASE;
+	$sql = "INSERT INTO series_id VALUES('".$code."','" . mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR)) . "','".$SERIES_DATABASE."')";
+	mysql_query ($sql) or die('1.Erreur SQL !'.$sql.'<br>'.mysql_error());
+}
+
+
+function banner_serie($dir){
+    global $DELETED_WORDS, $SERIE_WITH_BANNER, $SERIES_DIR;
+    if ($SERIE_WITH_BANNER == FALSE)
+        return;
+    $serieDir = explode('/', rtrim($SERIES_DIR, DIRECTORY_SEPARATOR));
+	$infos = explode('/', rtrim($dir, DIRECTORY_SEPARATOR));
+    $infos = array_reverse($infos);
+    if (count($infos) - count($serieDir) == 1)
+        {
+        // try to get serie code from DDB
+        $serie_infos = get_serie_code_infos($dir);
+        if (!empty($serie_infos)) {
+            $localBanner = "./images/poster_small/banner_".$serie_infos['id_serie'].".jpg";
+            if (file_exists($localBanner))
+                return $localBanner;
+            
+        } else
+        {
+            $name = $infos[0];
+            $allo = new AlloCine();
+            $recherche = $allo->serieSearch(keywordsAdapt($name,$DELETED_WORDS,1));//recherche serie
+	        if (!empty($recherche['code'])){
+	            $id=$recherche['code'];
+	            $serie = $allo->serieInfos($id);
+                set_serie_code($dir, $id);
+                $localBanner = "./images/poster_small/banner_".$id.".jpg";
+                if (!empty($serie['topBanner'])) {
+                    copy($serie['topBanner'],$localBanner);   
+                }
+	            return $serie['topBanner'];
+	        }
+        }
+    }
 }
 
 
@@ -384,7 +422,7 @@ function rep($rep){
             $dir = $rep;
 	    }
 	}
-    return $dir;
+    return $dir; 
 }
 
 function tri($sort){
@@ -451,7 +489,6 @@ function index_all() {
     $dirArray = array();
     $fileArray = array();
     $dirArray[] = $VIDEO_DIR;
-    debug("index_all");
     $sql = "SELECT link,dir FROM series WHERE 1 UNION SELECT link,dir FROM errors WHERE 1 UNION SELECT link,dir FROM movies WHERE 1" ;
     $req = mysql_query($sql) or die ('Erreur SQL : '.mysql_error());
     $base_movies = array();
@@ -505,17 +542,37 @@ function index_all() {
 }
 
 function index_auto($dir,$HIDDEN_FILES,$ext,$SERIES_DIR, $force_index = 0){
-//	if(is_serie($SERIES_DIR)) 
-    if(al_is_serie($dir)) 
-        $sql = "SELECT link FROM series WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' UNION SELECT link FROM errors WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' AND type='serie'";
-	else 
-        $sql = "SELECT link FROM movies WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' UNION SELECT link FROM errors WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' AND type='movie'";
-	$req = mysql_query($sql) or die ('Erreur SQL : '.mysql_error());
-	$base_movies = array();
+    
+    $base_movies = array();
 	$nonindexed = array();
-	while($data = mysql_fetch_array($req)){
-		$base_movies[] = stripslashes($data['link']);
-	}
+    if ($force_index == 1) {
+        if(al_is_serie($dir)) {
+            $sql =  "SELECT link FROM series WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' UNION SELECT link FROM errors WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' AND type='serie'";
+            $sql1 = "DELETE FROM series WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."'";
+            $sql2 = "DELETE FROM errors WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' AND type='serie'";
+        } else {
+            $sql = "SELECT link FROM movies WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' UNION SELECT link FROM errors WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' AND type='movie'";
+            $sql1 = "DELETE FROM movies WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."'";
+            $sql2 = "DELETE FROM errors WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' AND type='movie'";
+        }
+        try {
+            $req = mysql_query($sql1);
+            $req = mysql_query($sql2);
+        }catch (Exception $e) {
+            debug("SQL Exception : " . $e->getMessage());
+            logError("SQL Exception : " . $e->getMessage());
+            return;
+        }
+    } else {
+        if(al_is_serie($dir)) 
+            $sql = "SELECT link FROM series WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' UNION SELECT link FROM errors WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' AND type='serie'";
+    	else 
+            $sql = "SELECT link FROM movies WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' UNION SELECT link FROM errors WHERE dir='".mysql_real_escape_string(rtrim($dir, DIRECTORY_SEPARATOR))."' AND type='movie'";
+    	$req = mysql_query($sql) or die ('Erreur SQL : '.mysql_error());
+    	while($data = mysql_fetch_array($req)){
+    		$base_movies[] = stripslashes($data['link']);
+    	}
+    }
     $files = array();
 	if ($handle = opendir($dir)) {
 		while (false !== ($file = readdir($handle))) {
@@ -532,22 +589,22 @@ function index_auto($dir,$HIDDEN_FILES,$ext,$SERIES_DIR, $force_index = 0){
     	}
     closedir($handle);
 	}
-	else echo "Echec ouverture repertoire". $dir;
+	else {
+        debug ("Echec ouverture repertoire". $dir);
+        echo "Echec ouverture repertoire". $dir;
+	}
     
     
     
 	echo '<script>';
 	//echo '$("#empty").html("<div id=\"progressbar\" style=\"width:200px;\"></div>");';
 	echo "\n";
-    if ($force_index == 1) {
-        $tot = count($files);
-    echo 'var tabnonindexed = '.php2js($files).';';
-    } else {
-        $tot = count($nonindexed);
-        echo 'var tabnonindexed = '.php2js($nonindexed).';';
-    }
-    debug("auto_index - non indexed : " . $tot );
-    debug("auto_index - all files  : " . count($files) );
+    $tot = count($nonindexed);
+//    debug("auto_index - non indexed : " . $tot );
+//    debug("auto_index - all files  : " . count($files) );
+    
+    echo 'var tabnonindexed = '.php2js($nonindexed).';';
+
 	echo "\n";
 	echo 'if(tabnonindexed.length > 0){
 			$("#indexing").html("<br>'.indexing.'<br><br><div id=\"progressbar\" style=\"width:200px;\"></div><span id=\"nbindex\"></span><div id=\"error\"></div>").show();
@@ -585,23 +642,25 @@ function index_auto($dir,$HIDDEN_FILES,$ext,$SERIES_DIR, $force_index = 0){
 			
 	}); }';
 	echo '</script>';
-	$req = mysql_query($sql) or die ('Erreur SQL : '.mysql_error());
-	$nb_entree_bdd = mysql_num_rows($req); 
-    debug("auto_index nb_entree_bdd " . $nb_entree_bdd);
-    debug("auto_index count files " . count($files));
-	if ($nb_entree_bdd > count($files)){ //si le nb d'entree de la table mysql > nb entree effectif du rep
-		$i = 0;
-		while ($data = mysql_fetch_array($req)){
-			if (!in_array($data['link'],$files)){
-                debug("index_auto " . $data['link']);
-				mysql_query('DELETE FROM movies WHERE link="'.addslashes($data['link']).'"') or die ('Erreur SQL : '.mysql_error());
-				mysql_query("DELETE FROM movie_genre WHERE fk_id_movie = '".$data['id_movie']."'") or die ('Erreur SQL '.mysql_error());
-				echo addslashes($data['link']).' mis a jour<br>';
-				$i++;
-			}
-			if ($i == ($nb_entree_bdd-count($files))) break; //si on a atteint le nombre de fichier modifier on sort de la boucle
-		}	
-	}
+    if ($force_index == 0) {
+    	$req = mysql_query($sql) or die ('Erreur SQL : '.mysql_error());
+    	$nb_entree_bdd = mysql_num_rows($req); 
+//        debug("auto_index nb_entree_bdd " . $nb_entree_bdd);
+//        debug("auto_index count files " . count($files));
+    	if ($nb_entree_bdd > count($files)){ //si le nb d'entree de la table mysql > nb entree effectif du rep
+    		$i = 0;
+    		while ($data = mysql_fetch_array($req)){
+    			if (!in_array($data['link'],$files)){
+                    debug("index_auto 3 " . $data['link']);
+    				mysql_query('DELETE FROM movies WHERE link="'.addslashes($data['link']).'"') or die ('Erreur SQL : '.mysql_error());
+    				mysql_query("DELETE FROM movie_genre WHERE fk_id_movie = '".$data['id_movie']."'") or die ('Erreur SQL '.mysql_error());
+    				echo addslashes($data['link']).' mis a jour<br>';
+    				$i++;
+    			}
+    			if ($i == ($nb_entree_bdd-count($files))) break; //si on a atteint le nombre de fichier modifier on sort de la boucle
+    		}	
+    	}
+    }
 }
 
 function folders($dir,$HIDDEN_FILES){
